requirements.txt
python-dotenv==0.17.0
matplotlib==3.3.4
twitter==1.18.0
pymongo==3.11.3
pandas==1.2.4
wordcloud==1.8.1
numpy==1.20.1
nltk==3.5
textblob==0.15.3